feat(api): make WorkoutMapper

+++ backend/fit-journey-api/src/main/java/hu/hm/fitjourneyapi/repository/social/FriendRepository.java
@@ -3,5 +3,5 @@ package hu.hm.fitjourneyapi.repository.social;
 import hu.hm.fitjourneyapi.model.social.Friend;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FriendRepository extends JpaRepository<Friend, Integer> {
+public interface FriendRepository extends JpaRepository<Friend, Long> {
 }
